+++ src/main/java/kr/codesquad/LadderType.java
@@ -1,4 +1,4 @@
-package step1;
+package kr.codesquad;
 
 import java.util.stream.Stream;
