+++ src/main/java/kr/codesquad/Main.java
@@ -2,6 +2,7 @@ package kr.codesquad;
 
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Hello world!");
+        GameMaker gameMaker = new GameMaker();
+        gameMaker.gameStart();
     }
 }
